Move Predef traits to sub-package
